+++ packages/config/src/config.ts
@@ -1,510 +1,15 @@
-import type { WPTesterConfig, Tests, Environment } from "./wp-tester-config";
-import type { ResolvedWPTesterConfig, ResolvedEnvironment, ResolvedTests, ResolvedPHPUnitConfig, ResolvedBlueprint, ResolvedReporters, ResolvedPath } from "./resolved-types";
+import type { WPTesterConfig } from "./wp-tester-config";
+import type { ResolvedWPTesterConfig, ResolvedEnvironment, ResolvedTests, ResolvedBlueprint, ResolvedReporters, ResolvedPath } from "./resolved-types";
 import type { BlueprintV1Declaration } from "@wp-playground/blueprints";
-import { readFile, writeFile, access, constants as fsConstants } from "fs/promises";
-import { existsSync, statSync } from "fs";
-import { join, dirname, resolve, isAbsolute } from "path";
-import { fileURLToPath } from "url";
+import { readFile } from "fs/promises";
+import { existsSync } from "fs";
 import { getProjectRootMount } from "./auto-mount";
 import { detectProjectType } from "./options/project-type-detect";
-import { resolveBootstrapPath } from './options/phpunit-detect';
-import { hostToVfs } from './path-mappers';
-
-/**
- * Internal interface for expanded environments with resolved version arrays.
- * This is used internally during expansion before full resolution.
- */
-interface ExpandedEnvironment extends Environment {
-  /** The specific PHP version for this expanded environment */
-  _phpVersion?: string;
-  /** The specific WP version for this expanded environment */
-  _wpVersion?: string;
-}
+import { getProjectDir, resolveAbsolute, normalizeConfigPath } from "./path-utils";
+import { expandEnvironments } from "./environment-resolver";
+import { resolveTests } from "./test-resolver";
 
 export type { WPTesterConfig } from "./wp-tester-config";
-
-/**
- * Get the actual working directory, accounting for tools like tsx that may change process.cwd()
- * When running via npx or tsx, INIT_CWD contains the original directory where the command was invoked
- */
-export function getWorkingDirectory(): string {
-  return process.env.INIT_CWD || process.cwd();
-}
-
-/**
- * Resolve a path to an absolute path.
- * If the path is already absolute, return it as-is.
- * Otherwise, resolve it relative to the base directory.
- *
- * @param path - Path to resolve (relative or absolute)
- * @param baseDir - Base directory to resolve relative paths from
- * @returns Absolute path
- */
-export function resolveAbsolute(path: string, baseDir: string): string {
-  return isAbsolute(path) ? path : resolve(baseDir, path);
-}
-
-export function configPath(): string {
-  return join(getWorkingDirectory(), "wp-tester.json");
-}
-
-export function getSchemaPath(importMetaUrl?: string): string {
-  // For ESM in production (built files) - use import.meta.url from this module
-  // This works when the package is installed via npm or run via npx
-  const metaUrl = importMetaUrl || import.meta.url;
-  if (metaUrl) {
-    const currentDir = dirname(fileURLToPath(metaUrl));
-    const schemaPath = join(currentDir, "schema.json");
-    if (existsSync(schemaPath)) {
-      return schemaPath;
-    }
-  }
-
-  // For CommonJS (built files)
-  if (typeof __dirname !== "undefined") {
-    return join(__dirname, "schema.json");
-  }
-
-  // For dev mode: try to find the config package
-  // Could be in monorepo (packages/config) or installed (node_modules)
-  const candidates = [
-    // Monorepo from root
-    join(process.cwd(), "packages/config/src/schema.json"),
-    // Monorepo from a package directory
-    join(process.cwd(), "../config/src/schema.json"),
-    // Installed package
-    join(process.cwd(), "node_modules/@wp-tester/config/dist/schema.json"),
-  ];
-
-  for (const candidate of candidates) {
-    if (existsSync(candidate)) {
-      return candidate;
-    }
-  }
-
-  // Last resort
-  return candidates[candidates.length - 1];
-}
-
-export function getDefaultConfig(): WPTesterConfig {
-  return {
-    $schema:
-      "https://raw.githubusercontent.com/bgrgicak/wp-tester/trunk/packages/config/src/schema.json",
-    environments: [
-      {
-        name: "Latest WordPress and PHP",
-        blueprint: {
-          preferredVersions: {
-            php: "latest",
-            wp: "latest",
-          },
-        },
-      },
-    ],
-    tests: {},
-  };
-}
-
-export async function isConfigWritable(): Promise<boolean> {
-  try {
-    await access(configPath(), fsConstants.F_OK);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-export async function readConfigFile(path?: string): Promise<WPTesterConfig> {
-  if (!path) {
-    path = configPath();
-  }
-
-  // Normalize the path (handles directory paths)
-  path = normalizeConfigPath(path);
-
-  const content = await readFile(path, "utf8");
-  return JSON.parse(content) as WPTesterConfig;
-}
-
-export async function writeConfigFile(
-  config: WPTesterConfig,
-  path?: string
-): Promise<void> {
-  if (!path) {
-    path = configPath();
-  }
-  await writeFile(path, JSON.stringify(config, null, 2), "utf8");
-}
-
-/**
- * Get the absolute config file path.
- *
- * @param config - Path to config file (relative or absolute)
- * @returns Absolute path to the config file
- */
-export function getConfigPath(config: string): string {
-  // Expand tilde (~) to home directory
-  if (config.startsWith('~/') || config === '~') {
-    const homeDir = process.env.HOME || process.env.USERPROFILE;
-    if (!homeDir) {
-      throw new Error('Unable to resolve home directory');
-    }
-    config = config === '~' ? homeDir : join(homeDir, config.slice(2));
-  }
-
-  return isAbsolute(config) ? config : resolve(getWorkingDirectory(), config);
-}
-
-/**
- * Normalize a config path to ensure it points to a file, not a directory.
- * If the path is a directory, appends 'wp-tester.json' to it.
- * Used by readConfigFile, resolveConfig, and the test watcher to handle
- * directory paths passed via --config flag.
- *
- * @param configPath - Path to config file or directory (relative or absolute)
- * @returns Absolute path to the config file
- */
-export function normalizeConfigPath(configPath: string): string {
-  const absolutePath = getConfigPath(configPath);
-
-  // Check if path is a directory (synchronous)
-  try {
-    const stats = statSync(absolutePath);
-    if (stats.isDirectory()) {
-      return join(absolutePath, 'wp-tester.json');
-    }
-  } catch {
-    // If stat fails, assume it's a file path
-  }
-
-  return absolutePath;
-}
-
-/**
- * Get the config file directory path.
- *
- * @param configPath - Path to config file
- * @returns Directory path where the config file is located
- */
-export function getConfigDir(configPath: string): string {
-  return dirname(getConfigPath(configPath));
-}
-
-/**
- * Get the project root directory path.
- *
- * Returns config.projectHostPath if specified (resolved relative to config file or as absolute),
- * otherwise returns the config file's directory.
- *
- * @param config - Config object
- * @param configPath - Optional path to config file (needed to resolve relative projectHostPath)
- * @returns Absolute path to the project root directory
- */
-export function getProjectDir(
-  config: WPTesterConfig,
-  configPath?: string
-): string {
-  // Get base directory: config file location or cwd
-  const baseDir = configPath ? getConfigDir(configPath) : getWorkingDirectory();
-
-  // If projectHostPath is specified, resolve it relative to base directory
-  if (config.projectHostPath) {
-    return resolveAbsolute(config.projectHostPath, baseDir);
-  }
-
-  // No projectHostPath specified, return base directory
-  return baseDir;
-}
-
-/**
- * Normalize a version spec (string or array) to an array of versions.
- * @param spec - Version specification (single string or array)
- * @returns Array of version strings
- */
-function normalizeVersionSpec(spec: string | string[] | undefined): string[] {
-  if (!spec) {
-    return [];
-  }
-  return Array.isArray(spec) ? spec : [spec];
-}
-
-/**
- * Generate a name for an expanded environment based on the original name
- * and the specific PHP/WP versions.
- * @param baseName - Original environment name (if any)
- * @param phpVersion - PHP version for this combination (only if from matrix)
- * @param wpVersion - WordPress version for this combination (only if from matrix)
- * @param phpIsFromMatrix - Whether PHP version is from matrix expansion (not blueprint)
- * @param wpIsFromMatrix - Whether WP version is from matrix expansion (not blueprint)
- * @param isMatrix - Whether this is part of a matrix (multiple combinations)
- * @returns Generated environment name
- */
-function generateExpandedName(
-  baseName: string | undefined,
-  phpVersion: string | undefined,
-  wpVersion: string | undefined,
-  phpIsFromMatrix: boolean,
-  wpIsFromMatrix: boolean,
-  isMatrix: boolean
-): string | undefined {
-  // If not a matrix expansion, keep the original name
-  if (!isMatrix) {
-    return baseName;
-  }
-
-  // Build version suffix - only include versions that are from the matrix
-  const parts: string[] = [];
-  if (phpVersion && phpIsFromMatrix) {
-    parts.push(`PHP ${phpVersion}`);
-  }
-  if (wpVersion && wpIsFromMatrix) {
-    parts.push(`WP ${wpVersion}`);
-  }
-
-  if (parts.length === 0) {
-    return baseName;
-  }
-
-  const versionSuffix = parts.join(", ");
-  return baseName ? `${baseName} (${versionSuffix})` : versionSuffix;
-}
-
-/**
- * Expand environments with version arrays into multiple environments,
- * one for each combination of PHP and WP versions.
- *
- * Rules:
- * - If blueprint.preferredVersions.php is set, it overrides env.php (no matrix for PHP)
- * - If blueprint.preferredVersions.wp is set, it overrides env.wp (no matrix for WP)
- * - When arrays are provided for both php and wp, creates a full matrix of combinations
- *
- * @param environments - Original environments array
- * @param projectDir - Project directory for resolving blueprint paths
- * @returns Expanded environments array
- */
-async function expandEnvironments(
-  environments: Environment[],
-  projectDir: string
-): Promise<ExpandedEnvironment[]> {
-  const expanded: ExpandedEnvironment[] = [];
-
-  for (const env of environments) {
-    // Load blueprint if it's a string path (needed to check preferredVersions)
-    // If no blueprint is provided, use an empty object (defaults will be applied later)
-    let blueprint: BlueprintV1Declaration;
-    if (!env.blueprint) {
-      blueprint = {};
-    } else if (typeof env.blueprint === "string") {
-      const blueprintPath = resolveAbsolute(env.blueprint, projectDir);
-      const blueprintContent = await readFile(blueprintPath, "utf-8");
-      blueprint = JSON.parse(blueprintContent) as BlueprintV1Declaration;
-    } else {
-      blueprint = env.blueprint;
-    }
-
-    // Determine PHP versions to use
-    // Blueprint preferredVersions.php overrides environment-level php
-    // However, "latest" is treated as unspecified, allowing matrix expansion
-    const blueprintPhp = blueprint.preferredVersions?.php;
-    let phpVersions: string[];
-    let phpIsFromMatrix: boolean;
-    if (blueprintPhp && blueprintPhp !== "latest") {
-      // Blueprint takes precedence - use only the blueprint version
-      phpVersions = [blueprintPhp];
-      phpIsFromMatrix = false; // Not from matrix, from blueprint
-    } else {
-      // Use environment-level versions
-      phpVersions = normalizeVersionSpec(env.php);
-      phpIsFromMatrix = true; // From matrix (env-level config)
-    }
-
-    // Determine WP versions to use
-    // Blueprint preferredVersions.wp overrides environment-level wp
-    // However, "latest" is treated as unspecified, allowing matrix expansion
-    const blueprintWp = blueprint.preferredVersions?.wp;
-    let wpVersions: string[];
-    let wpIsFromMatrix: boolean;
-    if (blueprintWp && blueprintWp !== "latest") {
-      // Blueprint takes precedence - use only the blueprint version
-      wpVersions = [blueprintWp];
-      wpIsFromMatrix = false; // Not from matrix, from blueprint
-    } else {
-      // Use environment-level versions
-      wpVersions = normalizeVersionSpec(env.wp);
-      wpIsFromMatrix = true; // From matrix (env-level config)
-    }
-
-    // If no versions specified at either level, create a single environment
-    if (phpVersions.length === 0 && wpVersions.length === 0) {
-      expanded.push({ ...env });
-      continue;
-    }
-
-    // Ensure at least one version for each dimension (use empty string as placeholder for "not specified")
-    const phpList = phpVersions.length > 0 ? phpVersions : [""];
-    const wpList = wpVersions.length > 0 ? wpVersions : [""];
-
-    // Determine if this is a matrix (multiple combinations)
-    const isMatrix = phpList.length > 1 || wpList.length > 1;
-
-    // Create combinations
-    for (const phpVersion of phpList) {
-      for (const wpVersion of wpList) {
-        const expandedEnv: ExpandedEnvironment = {
-          ...env,
-          name: generateExpandedName(
-            env.name,
-            phpVersion || undefined,
-            wpVersion || undefined,
-            phpIsFromMatrix,
-            wpIsFromMatrix,
-            isMatrix
-          ),
-          _phpVersion: phpVersion || undefined,
-          _wpVersion: wpVersion || undefined,
-        };
-        expanded.push(expandedEnv);
-      }
-    }
-  }
-
-  return expanded;
-}
-
-/**
- * Convert a host path to a ResolvedPath using the project path mapping
- * @param hostPath - Absolute path on host
- * @param projectPath - Project path mapping
- * @returns ResolvedPath with both host and VFS paths
- */
-function toResolvedPath(hostPath: string, projectPath: ResolvedPath): ResolvedPath {
-  return {
-    hostPath,
-    vfsPath: hostToVfs(hostPath, projectPath),
-  };
-}
-
-/**
- * PHPUnit flags that are boolean (do not take a value).
- * If a flag is NOT in this list, we assume it takes a value,
- * and therefore the next argument should NOT be resolved as a path.
- */
-const PHPUNIT_BOOLEAN_FLAGS = new Set([
-  '--teamcity', '--debug', '--verbose', '--testdox',
-  '--stderr', '--stop-on-error', '--stop-on-failure', '--stop-on-warning',
-  '--stop-on-defect', '--stop-on-risky', '--stop-on-skipped', '--stop-on-incomplete',
-  '--stop-on-deprecation', '--stop-on-notice',
-  '--fail-on-warning', '--fail-on-risky', '--fail-on-incomplete', '--fail-on-skipped',
-  '--fail-on-deprecation', '--fail-on-notice',
-  '--strict-coverage', '--strict-global-state', '--disallow-test-output',
-  '--disallow-resource-usage', '--enforce-time-limit',
-  '--process-isolation', '--no-globals-backup', '--static-backup',
-  '--no-configuration', '--no-coverage', '--no-logging', '--no-interaction',
-  '--no-extensions', '--no-output', '--dont-report-useless-tests',
-  '--no-progress',
-  '--display-deprecations', '--display-errors', '--display-incomplete',
-  '--display-notices', '--display-skipped', '--display-warnings',
-  '--strict',
-  '--help', '--version',
-  '--list-groups', '--list-suites', '--list-tests', '--list-tests-xml',
-]);
-
-/**
- * Resolve PHPUnit arguments to map host paths to VFS paths
- *
- * @param args - Original arguments from config
- * @param projectPath - Project path with host and VFS paths
- * @returns Resolved arguments with paths converted to VFS paths
- */
-export function resolvePhpunitArgs(args: string[], projectPath: ResolvedPath): string[] {
-  const resolvedArgs: string[] = [];
-
-  for (let i = 0; i < args.length; i++) {
-    const arg = args[i];
-
-    // Don't resolve flags
-    if (arg.startsWith('-')) {
-      resolvedArgs.push(arg);
-      continue;
-    }
-
-    // Check if the previous argument was a flag that takes a value
-    const prevArg = args[i - 1];
-    if (prevArg?.startsWith('-') && !prevArg.includes('=')) {
-      // If the previous flag is NOT a known boolean flag, assume it takes a value
-      const flagName = prevArg.split('=')[0];
-      if (!PHPUNIT_BOOLEAN_FLAGS.has(flagName)) {
-        resolvedArgs.push(arg); // This is the flag's value, not a path
-        continue;
-      }
-    }
-
-    // Resolve arguments that look like file/directory paths
-    const looksLikePath = arg.includes('/') ||
-                          arg.endsWith('.php') ||
-                          arg === 'tests' || arg === 'test';
-
-    if (looksLikePath) {
-      const absoluteHostPath = resolveAbsolute(arg, projectPath.hostPath);
-      const vfsPath = hostToVfs(absoluteHostPath, projectPath);
-      resolvedArgs.push(vfsPath);
-      continue;
-    }
-
-    resolvedArgs.push(arg);
-  }
-
-  return resolvedArgs;
-}
-
-/**
- * Resolve PHPUnit config paths and set default testMode
- * @param tests - Tests configuration
- * @param projectPath - Project path with host and VFS paths
- * @returns Resolved tests configuration
- */
-async function resolveTests(tests: Tests, projectPath: ResolvedPath): Promise<ResolvedTests> {
-  // If no PHPUnit config, return tests as-is (but explicitly typed)
-  if (!tests.phpunit) {
-    return {
-      plugin: tests.plugin,
-      theme: tests.theme,
-      wp: tests.wp,
-      passWithNoTests: tests.passWithNoTests,
-    };
-  }
-
-  // Resolve PHPUnit config with absolute paths and default testMode
-  const phpunit = tests.phpunit;
-  const resolvedConfigHostPath = resolveAbsolute(phpunit.configPath, projectPath.hostPath);
-
-  // Resolve bootstrap path using centralized logic
-  const bootstrapHostPath = await resolveBootstrapPath(
-    resolvedConfigHostPath,
-    projectPath.hostPath,
-    phpunit.bootstrapPath
-  );
-
-  const resolvedPhpunit: ResolvedPHPUnitConfig = {
-    phpunitPath: toResolvedPath(resolveAbsolute(phpunit.phpunitPath, projectPath.hostPath), projectPath),
-    configPath: toResolvedPath(resolvedConfigHostPath, projectPath),
-    testMode: phpunit.testMode ?? "unit",
-    bootstrapPath: bootstrapHostPath ? toResolvedPath(bootstrapHostPath, projectPath) : undefined,
-  };
-
-  // Resolve phpunitArgs paths if present
-  if (phpunit.phpunitArgs) {
-    resolvedPhpunit.phpunitArgs = resolvePhpunitArgs(phpunit.phpunitArgs, projectPath);
-  }
-
-  return {
-    plugin: tests.plugin,
-    theme: tests.theme,
-    wp: tests.wp,
-    phpunit: resolvedPhpunit,
-    passWithNoTests: tests.passWithNoTests,
-  };
-}
 
 /**
  * Resolve config from path or object and adjust relative paths to absolute paths
